Move Startup scenario to dedicated folder.
To set an example for isolated cucumber tests.
