+++ src/test/java/nl/tudelft/jpacman/e2e/framework/startup/StartupTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.jpacman.e2e.framework;
+package nl.tudelft.jpacman.e2e.framework.startup;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.SnippetType;
@@ -14,9 +14,9 @@ import org.junit.runner.RunWith;
 @CucumberOptions(
 	plugin = {"pretty"}, 
 	snippets = SnippetType.CAMELCASE,
-	glue = {"nl.tudelft.jpacman.e2e.framework"},
-	features = "classpath:frameworkfeatures")
-public class CucumberTest {
+	glue = {"nl.tudelft.jpacman.e2e.framework.startup"},
+	features = "classpath:frameworkfeatures/startup")
+public class StartupTest {
 
 	/**
 	 * This class should be empty, step definitions should be in separate classes.
